Renamed project, folders, packages and classes to explainingHttpServlet. Removed all comments except from *.java files Removed @ symbol where not required in JavaDocs comments Updated project names in parent pom files. Signed-off-by:Ken Fogel <omniprof@gmail.com>
