+++ servlet/README.md
@@ -12,4 +12,4 @@ upon.
 5. [A ServletContextListener example](servletContextListener/README.md)
 6. [A ServletRequestListener example](servletRequestListener/README.md)
 7. [A @WebServlet example](webServlet/README.md)
-8. [The Learning Servlet](theLearningServlet/README.md)
+8. [A Explaining HttpServlet](explainingHttpServlet/README.md)
