+++ servlet/explainingHttpServlet/README.md
@@ -1,0 +1,3 @@
+# Explaining HttpServlet example
+
+This example demonstrates the use of the four doXXX methods commonly used in a Servlet class. The doGet and doPost methods can be accessed from an HTML form. The doPut and doDelete methods cannot. You can access these two methods with the tool cURL and this is explained when you run the application. They are in the example because these methods along with doGet and doPost form the basis for RESTful web services.
