+++ servlet/explainingHttpServlet/pom.xml
@@ -1,0 +1,36 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>jakartaee.examples.servlet</groupId>
+    <artifactId>explainingHttpServlet</artifactId>
+    <version>8-SNAPSHOT</version>
+    <packaging>war</packaging>
+    <name>A Servlet API Explaining HttpServlet</name>
+    <description>An example servlet for the Jakarta EE Examples project</description>
+
+    <developers>
+        <developer>
+            <name>Ken Fogel</name>
+            <email>omniprof@gmail.com</email>
+        </developer>
+    </developers>
+
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+    
+    <parent>
+        <artifactId>project</artifactId>
+        <groupId>jakartaee.examples.servlet</groupId>
+        <version>8-SNAPSHOT</version>
+    </parent>
+
+    <dependencies>
+        <dependency>
+            <groupId>javax</groupId>
+            <artifactId>javaee-web-api</artifactId>
+           <scope>provided</scope>
+        </dependency>
+    </dependencies>
+
+</project>
