+++ servlet/explainingHttpServlet/src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,6 @@
+<web-app version="4.0" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_4_0.xsd">    <!-- Name displayed in Server Console -->
+    <display-name>ExplainingHttpServlet</display-name>
+    <welcome-file-list>
+        <welcome-file>index.html</welcome-file>
+    </welcome-file-list>
+</web-app>
