+++ servlet/explainingHttpServlet/src/main/webapp/index.html
@@ -1,13 +1,5 @@
 <!DOCTYPE html>
-<!-- TheLearningServlet Web Application
-     Prepared by Ken Fogel omniprof@gmail.com
-     Twitter: @omniprof
--->
 <html>
-
-    <!-- This is a basic, old style, html file that allows you to test GET & POST
-         and shows you the curl command to test PUT & DELETE.
-    -->
     <head>
         <link rel='stylesheet' type='text/css' href='styles/main.css'>
         <title>The Learning Servlet</title>
@@ -34,11 +26,11 @@
         <br/>
         <label>PUT method</label>
         <br/>
-        <code>curl -X PUT http://localhost:8080/TheLearningServlet/learning</code>
+        <code>curl -X PUT http://localhost:8080/explainingHttpServlet/learning</code>
         <br/>
         <br/>
         <label>DELETE method</label>
         <br/>
-        <code>curl -X DELETE http://localhost:8080/TheLearningServlet/learning</code>
+        <code>curl -X DELETE http://localhost:8080/explainingHttpServlet/learning</code>
     </body>
 </html>
