+++ servlet/explainingHttpServlet/src/main/webapp/styles/main.css
@@ -1,8 +1,3 @@
-/* TheLearningServlet Web Application CSS
-     Prepared by Ken Fogel
-     Twitter: @omniprof
-*/
-
 label {
     float: left;
     /* width: 7em; */
