--- servlet/theLearningServlet/README.md
@@ -1,4 +1,0 @@
-# The Learning Servlet example
-
-This example demonstrates the use of the four doXXX methods commonly used in a Servlet class.
-
